Fix a warning in bitmap checksummer.
Review URL: https://codereview.chromium.org/12629003

git-svn-id: http://skia.googlecode.com/svn/trunk@8022 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/utils/SkBitmapChecksummer.cpp
@@ -25,8 +25,8 @@ static void write_int_to_buffer(int val, char* buf) {
 
 /*static*/ uint64_t SkBitmapChecksummer::Compute64Internal(
         const SkBitmap& bitmap, const SkBitmapTransformer& transformer) {
-    int pixelBufferSize = transformer.bytesNeededTotal();
-    int totalBufferSize = pixelBufferSize + 8; // leave room for x/y dimensions
+    size_t pixelBufferSize = transformer.bytesNeededTotal();
+    size_t totalBufferSize = pixelBufferSize + 8; // leave room for x/y dimensions
 
     SkAutoMalloc bufferManager(totalBufferSize);
     char *bufferStart = static_cast<char *>(bufferManager.get());
